fix vis_keypoint

+++ tests/visualizations_tests/test_vis_keypoint.py
@@ -21,7 +21,7 @@ class TestVisKeypoint(unittest.TestCase):
 
     def test_vis_keypoint(self):
         if optional_modules:
-            img = np.random.uniform(size=(3, 32, 32)).astype(np.uint8)
+            img = np.random.uniform(size=(3, 32, 32)).astype(np.float32)
             keypoint = np.random.uniform(size=(3, 2)).astype(np.float32)
             ax = vis_keypoint(img, keypoint, self.kp_mask)
